refactor(components/selector4nix-streaming): rename StreamImpl to FullStream
