+++ components/selector4nix-streaming/src/stream/mod.rs
@@ -1,0 +1,3 @@
+mod full;
+
+pub use full::FullStream;
