+++ components/selector4nix-streaming/src/throttler/mod.rs
@@ -1,0 +1,8 @@
+mod per_host;
+
+pub use per_host::PerHostHttpThrottler;
+
+use tokio::sync::OwnedSemaphorePermit;
+
+#[expect(dead_code)]
+pub struct ThrottlerPermit(OwnedSemaphorePermit);
